+++ cli-tools/cli/bin/aofl.js
@@ -22,7 +22,8 @@ aofl source
   .command('dom-scope [path]', 'validate generated dom-scope')
   .command('build', 'Builds your AofL JS project.')
   .command('test', 'Runs unit tests')
-  .command('serve', 'Builds and servers your AofL JS project.');
+  .command('serve', 'Builds and servers your AofL JS project.')
+  .command('upgrade <path>', 'Upgrade your AofL JS project to the next version.');
 
 program
   .parse(process.argv);