+++ cli-tools/cli/commands/upgrade-project/codemod/2.x/04-add-jsdocrc-config/index.js
@@ -1,0 +1,27 @@
+const chalk = require('chalk');
+const path = require('path');
+const fs = require('fs');
+
+module.exports = (projectRoot, projectInfo, commit = false) => {
+  const configPath = path.join(projectRoot, '.jsdocrc.json');
+  const templateFilePath = path.join(__dirname, 'templates', '.jsdocrc.json.sample');
+  const template = fs.readFileSync(templateFilePath, {encoding: 'utf-8'});
+
+  const result = {
+    skip: true,
+    message: chalk.yellow(`${configPath} will be generated`),
+    label: chalk.cyan(`Generating ${configPath}`)
+  };
+
+  try {
+    fs.lstatSync(configPath);
+  } catch (e) {
+    result.skip = false;
+  }
+
+  if (commit) {
+    fs.writeFileSync(configPath, template, {encoding: 'utf-8'});
+  }
+
+  return result;
+};