+++ cli-tools/cli/commands/upgrade-project/codemod/2.x/10-transform-routesconfig/transform.js
@@ -1,0 +1,24 @@
+module.exports = (file, api, options) => {
+  const {j} = api;
+
+  const root = j(file.source);
+  const routesImport = root.find(j.ImportDeclaration, {
+    source: {
+      type: 'Literal',
+      value: '@aofl/templating-plugin/routes.config',
+    },
+  });
+
+  let defaultVarName = 'routesConfig';
+
+  routesImport.find(j.ImportDefaultSpecifier).forEach((x) => {
+    defaultVarName = x.node.local.name;
+  });
+
+  const newImport = j.importDeclaration([j.importDefaultSpecifier(j.identifier(defaultVarName))], j.literal('./__config/routes'));
+
+  routesImport.insertAfter(newImport);
+  routesImport.remove();
+
+  return root.toSource({quote: 'single', objectCurlySpacing: false});
+};