+++ cli-tools/cli/commands/upgrade-project/upgrade-config.js
@@ -1,0 +1,7 @@
+module.exports = [
+  {
+    from: '1.x',
+    to: 'latest v2',
+    runner: require('./codemod/2.x')
+  }
+];