+++ cli-tools/cli/lib/constants-enumerate.js
@@ -1,0 +1,8 @@
+const project = {
+  GENERATED_CONFIG_FILE_NAMES: ['.aofl.json'],
+  CONFIG_FILE_NAMES: ['.aofl.js'],
+};
+
+module.exports = {
+  project
+};