+++ aofl-js-packages/web-components/aofl-element/index.js
@@ -1,2 +1,3 @@
 export {AoflElement} from './modules/aofl-element';
 export * from 'lit-element';
+export {html} from 'lit-html';